Initial DynamoRecord schema

+++ dynamo_query/dynamo_table.py
@@ -8,15 +8,13 @@ from typing import (
     Iterable,
     Iterator,
     List,
-    Mapping,
     Optional,
     Set,
     Type,
     TypeVar,
+    Union,
     cast,
 )
-
-from typing_extensions import Literal
 
 from dynamo_query import json_tools
 from dynamo_query.data_table import DataTable
@@ -25,6 +23,9 @@ from dynamo_query.dynamo_query_types import (
     AttributeDefinitionTypeDef,
     CreateTableOutputTypeDef,
     DynamoDBClient,
+    PartitionKeyOperatorTypeDef,
+    ProvisionedThroughputTypeDef,
+    SortKeyOperatorTypeDef,
     Table,
 )
 from dynamo_query.dynamo_record import DynamoRecord
@@ -35,7 +36,7 @@ from dynamo_query.utils import chunkify
 
 __all__ = ("DynamoTable", "DynamoTableError")
 
-_R = TypeVar("_R", bound=Mapping[str, Any])
+_R = TypeVar("_R", bound=DynamoRecord)
 
 
 class DynamoTableError(BaseException):
@@ -137,7 +138,14 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
     # class to use as DynamoQuery for easier subclassing
     dynamo_query_class: Type[DynamoQuery] = DynamoQuery
 
-    record_class: Optional[DynamoRecord] = None
+    # Custom record class inherited from DynamoRecord
+    record_class: Optional[Type[DynamoRecord]] = None
+
+    # ProvisionedThroughput parameters
+    provisioned_throughput: Optional[ProvisionedThroughputTypeDef] = {
+        "ReadCapacityUnits": 50,
+        "WriteCapacityUnits": 10,
+    }
 
     def __init__(self, logger: Optional[logging.Logger] = None):
         self._lazy_logger = logger
@@ -182,9 +190,34 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
 
         return self.get_sort_key(record)
 
+    def _convert_record(self, record: Union[_R, Dict]) -> _R:
+        # pylint: disable=isinstance-second-argument-not-valid-type
+        if self.record_class and not isinstance(record, self.record_class):
+            # pylint: disable=not-callable
+            return self.record_class(record)  # type: ignore
+
+        return record  # type: ignore
+
+    def get_table_status(self) -> Optional[str]:
+        """
+        Get table status from Dynamo.
+
+        Returns:
+            Status string or None.
+        """
+        try:
+            response = self.client.describe_table(TableName=self.table.name)
+        except self.client.exceptions.ResourceNotFoundException:
+            return None
+
+        return response["Table"]["TableStatus"]
+
     def delete_table(self) -> None:
         """
         Delete the table from DynamoDB.
+
+        If table is creating, wait until it is created, then deletes it.
+        If table is deleting or does not exist, does nothing.
 
         Example:
 
@@ -194,11 +227,30 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
 
             # delete table
             UserTable.delete_table()
+
+            # make sure that it is deleted
+            user_table.wait_until_not_exists()
             ```
         """
+        status = self.get_table_status()
+        self._logger.debug(f"Table {self.table.name} status is {status}")
+
+        if status is None:
+            self._logger.debug(f"Table {self.table.name} does not exist, skipping deletion")
+            return
+
+        if status == "DELETING":
+            self._logger.debug(f"Table {self.table.name} is deleting, skipping deletion")
+            return
+
+        if status == "CREATING":
+            self._logger.debug(f"Table {self.table.name} is creating, waiting until created")
+            self.wait_until_exists()
+
+        self._logger.debug(f"Deleting {self.table.name}")
         self.table.delete()
 
-    def create_table(self) -> CreateTableOutputTypeDef:
+    def create_table(self) -> Optional[CreateTableOutputTypeDef]:
         """
         Create a table in DynamoDB.
 
@@ -212,19 +264,44 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             UserTable.create_table()
             ```
         """
+        status = self.get_table_status()
+        self._logger.debug(f"Table {self.table.name} status is {status}")
+
+        if status == "DELETING":
+            self._logger.debug(f"Table {self.table.name} is deleting, waiting until deleted")
+            self.wait_until_not_exists()
+
+        if status == "CREATING":
+            self._logger.debug(f"Table {self.table.name} is creating, skipping")
+            return None
+
+        if status == "ACTIVE":
+            self._logger.debug(f"Table {self.table.name} is active, skipping")
+            return None
+
         global_secondary_indexes = [
-            i.as_global_secondary_index() for i in self.global_secondary_indexes
+            i.as_global_secondary_index(self.provisioned_throughput)
+            for i in self.global_secondary_indexes
         ]
         local_secondary_indexes = [
             i.as_local_secondary_index() for i in self.local_secondary_indexes
         ]
 
+        extra_params: Dict[str, Any] = {}
+
+        if global_secondary_indexes:
+            extra_params["GlobalSecondaryIndexes"] = global_secondary_indexes
+        if local_secondary_indexes:
+            extra_params["LocalSecondaryIndexes"] = local_secondary_indexes
+
+        if self.provisioned_throughput:
+            extra_params["ProvisionedThroughput"] = self.provisioned_throughput
+
         return self.client.create_table(
             AttributeDefinitions=self._attribute_definitions,
             TableName=self.table.name,
             KeySchema=self.primary_index.as_key_schema(),
-            GlobalSecondaryIndexes=global_secondary_indexes,
-            LocalSecondaryIndexes=local_secondary_indexes,
+            **extra_params,
         )
 
     def _get_attribute_definitions(self,) -> List[AttributeDefinitionTypeDef]:
@@ -284,7 +361,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             value = record[attribute_name]
             if not isinstance(value, attribute_type):
                 raise DynamoTableError(
-                    f"Attribute {attribute_name} has invalid type, {attribute_type} expected.",
+                    f"Attribute {attribute_name} has invalid type, {attribute_type} expected, got {type(value)}.",
                     data=record,
                 )
 
@@ -353,7 +430,9 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             records = self.scan(projection=self.table_keys)
 
         for records_chunk in chunkify(records, self.max_batch_size):
-            existing_records = DataTable[_R]().add_record(*records_chunk)
+            existing_records = DataTable[_R](record_class=self.record_class).add_record(
+                *records_chunk
+            )
             self.dynamo_query_class.build_batch_delete_item(logger=self._logger).table(
                 table_keys=self.table_keys, table=self.table,
             ).execute(existing_records)
@@ -395,10 +474,12 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         Returns:
             DataTable with existing records.
         """
+
         if not data_table:
-            return DataTable()
+            return DataTable[_R](record_class=self.record_class)
         get_data_table = DataTable[_R]()
         for record in data_table.get_records():
+            record = self._convert_record(record)
             get_data_table.add_record(
                 {
                     **record,
@@ -407,7 +488,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
                 }
             )
 
-        results: DataTable[_R] = DynamoQuery.build_batch_get_item(logger=self._logger,).table(
+        results: DataTable[_R] = DynamoQuery.build_batch_get_item(logger=self._logger).table(
             table_keys=self.table_keys, table=self.table
         ).execute(data_table=get_data_table)
         return results
@@ -521,27 +602,27 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
 
         update_data_table: DataTable[_R] = DataTable()
         for record_index, record in enumerate(existing_records.get_records()):
-            updated_record = data_table.get_record(record_index)
+            updated_record = self._convert_record(data_table.get_record(record_index))
             preserve_keys_record: Dict[str, Any] = {}
             for key in set_if_not_exists:
                 if key in record:
                     preserve_keys_record[key] = record[key]
 
-            new_record = cast(
-                _R,
+            new_record = self._convert_record(
                 {
                     **record,
                     **updated_record,
                     **preserve_keys_record,
                     "dt_created": record.get("dt_created") or now_str,
                     "dt_modified": now_str,
-                },
+                }
             )
+
             normalized_record = self.normalize_record(new_record)
             self.validate_record_attributes(normalized_record)
-            update_data_table.add_record(normalized_record)
+            update_data_table.add_record(dict(normalized_record))
 
-        results: DataTable[_R] = DynamoQuery.build_batch_update_item(logger=self._logger,).table(
+        results: DataTable[_R] = DynamoQuery.build_batch_update_item(logger=self._logger).table(
             table_keys=self.table_keys, table=self.table,
         ).execute(update_data_table)
         return results
@@ -562,7 +643,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             get_data_table = DataTable[_R]().add_record(*records_chunk)
             result_data_table = self.batch_get(get_data_table)
             for record in result_data_table.get_records():
-                yield record
+                yield self._convert_record(record)
 
     def batch_delete_records(self, records: Iterable[_R]) -> Iterator[_R]:
         """
@@ -577,10 +658,13 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             Deleted or not found record data.
         """
         for records_chunk in chunkify(records, self.max_batch_size):
-            upsert_data_table = DataTable[_R]().add_record(*records_chunk)
-            result = self.batch_delete(upsert_data_table)
+            delete_data_table = DataTable[_R](record_class=self.record_class).add_record(
+                *records_chunk
+            )
+            result = self.batch_delete(delete_data_table)
             for record in result.get_records():
-                yield record
+                print(record)
+                yield self._convert_record(record)
 
     def batch_upsert_records(
         self, records: Iterable[_R], set_if_not_exists_keys: Iterable[str] = (),
@@ -598,12 +682,14 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             Created, updated or not found record data.
         """
         for records_chunk in chunkify(records, self.max_batch_size):
-            upsert_data_table = DataTable[_R]().add_record(*records_chunk)
+            upsert_data_table = DataTable[_R](record_class=self.record_class).add_record(
+                *records_chunk
+            )
             result = self.batch_upsert(
                 upsert_data_table, set_if_not_exists_keys=set_if_not_exists_keys
             )
             for record in result.get_records():
-                yield record
+                yield self._convert_record(record)
 
     def get_record(self, record: _R) -> Optional[_R]:
         """
@@ -637,6 +723,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         Returns:
             A dict with record data or None.
         """
+        record = self._convert_record(record)
         partition_key = self._get_partition_key(record)
         sort_key = self._get_sort_key(record)
         result = (
@@ -647,7 +734,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         if set(result.get_set_column_names()).issubset(self.table_keys):
             return None
 
-        return cast(_R, result.get_record(0))
+        return self._convert_record(result.get_record(0))
 
     def upsert_record(
         self,
@@ -703,8 +790,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         now = datetime.datetime.utcnow()
         now_str = now.isoformat()
 
-        new_record = cast(
-            _R,
+        new_record = self._convert_record(
             {
                 self.partition_key_name: partition_key,
                 self.sort_key_name: sort_key,
@@ -712,8 +798,9 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
                 **(extra_data or {}),
                 "dt_modified": now_str,
                 "dt_created": now_str,
-            },
+            }
         )
+
         new_record = self.normalize_record(new_record)
         update_keys = set(new_record.keys()) - self.table_keys - set_if_not_exists
         update_keys.add("dt_modified")
@@ -725,7 +812,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             .table(table_keys=self.table_keys, table=self.table,)
             .execute_dict(cast(Dict[str, Any], new_record))
         )
-        return result.get_record(0)
+        return self._convert_record(result.get_record(0))
 
     def delete_record(
         self, record: _R, condition_expression: Optional[ConditionExpression] = None,
@@ -770,7 +857,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         )
         if not result:
             return None
-        return result.get_record(0)
+        return self._convert_record(result.get_record(0))
 
     def scan(
         self,
@@ -827,13 +914,13 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             query_data.update(data)
 
         for record in self._yield_from_query(query, data=query_data, limit=limit):
-            yield record
+            yield self._convert_record(record)
 
     def query(
         self,
-        partition_key: str,
+        partition_key: Any,
         index: DynamoTableIndex = primary_index,
-        sort_key: Optional[str] = None,
+        sort_key: Optional[Any] = None,
         sort_key_prefix: Optional[str] = None,
         filter_expression: Optional[ConditionExpressionType] = None,
         scan_index_forward: bool = True,
@@ -891,8 +978,8 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
         Yields:
             Matching record.
         """
-        sort_key_operator: Literal["=", "begins_with"] = "="
-        partition_key_operator: Literal["="] = "="
+        sort_key_operator: SortKeyOperatorTypeDef = "="
+        partition_key_operator: PartitionKeyOperatorTypeDef = "="
         if sort_key_prefix is not None:
             sort_key_operator = "begins_with"
             sort_key = sort_key_prefix
@@ -925,7 +1012,7 @@ class DynamoTable(Generic[_R], LazyLogger, ABC):
             query_data.update(data)
 
         for record in self._yield_from_query(query, data=query_data, limit=limit):
-            yield record
+            yield self._convert_record(record)
 
     def wait_until_exists(self) -> None:
         """
